requirements.txt

gradio>=4.40.0
python-dotenv>=1.0.1
requests>=2.32.3
